% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cate_list_url_ex}
\alias{cate_list_url_ex}
\title{Category Page Url Example}
\format{
example url character
}
\usage{
cate_list_url_ex
}
\description{
Category Page Url Example
}
\keyword{datasets}
