% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{checkContentSummary}
\alias{checkContentSummary}
\title{Check Content Summary.}
\usage{
checkContentSummary(html_obj,
  summary_node_info = "div.media_end_head_autosummary", summary_attr = "")
}
\arguments{
\item{html_obj}{"xml_document" "xml_node" using read_html function.}

\item{summary_node_info}{Information about node names like tag with class or id. Default is "div.article_info h3" for naver news title.}

\item{summary_attr}{if you want to get attribution text, please write down here.}
}
\value{
Get character summary content.
}
\description{
Get naver news summary from link.
}
