% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{Deprecated-N2H4}
\alias{Deprecated-N2H4}
\alias{getVideoUrl}
\title{Deprecated functions in the N2H4 package}
\usage{
getVideoUrl(...)
}
\arguments{
\item{...}{arguments to be passed to the functions
(obscured to enforce the usage of new functions)}
}
\description{
Deprecated functions in the N2H4 package
}
\keyword{internal}
