% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComment.R
\name{getAllComment}
\alias{getAllComment}
\title{Get All Comment}
\usage{
getAllComment(turl)
}
\arguments{
\item{turl}{character. News article on 'Naver' such as
<https://n.news.naver.com/mnews/article/023/0003712918>.
News article url that is not on Naver.com domain will generate an error.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Get all comments from the provided news article url on naver
}
\details{
Works just like getComment, but this function executed in a fashion
  where it finds and extracts all comments from the given url.
}
\examples{
\dontrun{
  getAllComment("https://n.news.naver.com/mnews/article/214/0001195110")
  }
}
