% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommentHistory.R
\name{getAllCommentHistory}
\alias{getAllCommentHistory}
\title{Get All Comment History}
\usage{
getAllCommentHistory(turl, commentNo)
}
\arguments{
\item{turl}{character. News article on 'Naver' such as
<https://n.news.naver.com/mnews/article/001/0009205077?sid=102>.
News articl url that is not on Naver.com domain will generate an error.}

\item{commentNo}{Parent Comment No.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Get All Comment History
}
\examples{
\dontrun{
  getAllComment("https://n.news.naver.com/mnews/article/214/0001195110?sid=103")
  }
}
