% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllUrlList.R
\name{getAllUrlListByCategory}
\alias{getAllUrlListByCategory}
\title{Get naver news titles and links from target url.}
\usage{
getAllUrlListByCategory(urls = urls)
}
\arguments{
\item{url}{is data.frame generated from setUrl function.}
}
\value{
Get data.frame(news_title, news_links).
}
\description{
Get naver news titles and links from target url.
}

