% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllUrlList.R
\name{getAllUrlListByQuery}
\alias{getAllUrlListByQuery}
\title{Get naver news titles and links from target url.}
\usage{
getAllUrlListByQuery(urls = urls)
}
\arguments{
\item{urls}{is data.frame generated from setUrl function.}
}
\value{
Get data.frame(news_title, news_links).
}
\description{
Get naver news titles and links from target url.
}

