% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCategory.R
\name{getCategory}
\alias{getCategory}
\alias{news_category_get}
\title{News Category}
\usage{
getCategory(fresh = FALSE)

news_category_get(fresh = FALSE)
}
\arguments{
\item{fresh}{get data from online. Default is FALSE using cached built-in data.}
}
\description{
News Category
}
