% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCategoryUrl.R
\name{getCategoryUrl}
\alias{getCategoryUrl}
\title{Get naver news sub_category names and urls recently.}
\usage{
getCategoryUrl()
}
\value{
Get data.frame(cate_name, cate_url).
}
\description{
No pram needed.
Main category urls are written manually like blow.
cate_code<-c(100,101,102,103,104,105)
}

