% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getContentBody}
\alias{getContentBody}
\title{Get Content body name.}
\usage{
getContentBody(html_obj, node_info = "div#articleBodyContents", attr = "")
}
\arguments{
\item{html_obj}{"xml_document" "xml_node" using read_html function.}

\item{node_info}{Information about node names like tag with class or id. Default is "div.article_info h3" for naver news title.}

\item{attr}{if you want to get attribution text, please write down here.}
}
\value{
Get character body content.
}
\description{
Get naver news body from link.
}
