% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getContentDatetime}
\alias{getContentDatetime}
\title{Get Content datetime}
\usage{
getContentDatetime(
  html_obj,
  datetime_node_info = "span.t11",
  datetime_attr = "",
  getEdittime = TRUE
)
}
\arguments{
\item{html_obj}{"xml_document" "xml_node" using read_html function.}

\item{datetime_node_info}{Information about node names like tag with class or id. Default is "div.article_info h3" for naver news title.}

\item{datetime_attr}{if you want to get attribution text, please write down here.}

\item{getEdittime}{if TRUE, can get POSIXlt type datetime length 2 means published time and final edited time. if FALSE, get Date length 1.}
}
\value{
Get POSIXlt type datetime.
}
\description{
Get naver news published datetime from link.
}
\examples{
\donttest{
  print(news_url_ex)
  hobj <- xml2::read_html(news_url_ex)
  getContentDatetime(hobj)
  }
}
