% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContentParallel.R
\name{getContentParallel}
\alias{getContentParallel}
\title{Get naver news content from links parallelly.}
\usage{
getContentParallel(url = url)
}
\arguments{
\item{url}{is naver news links.}
}
\value{
Get data.frame(url,datetime,press,title,content).
}
\description{
Get naver news content from links parallelly.
}

