% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getContentPress}
\alias{getContentPress}
\title{Get Content Press name.}
\usage{
getContentPress(html_obj, node_info = "div.article_header div a img",
  attr = "title")
}
\arguments{
\item{html_obj}{"xml_document" "xml_node" using read_html function.}

\item{node_info}{Information about node names like tag with class or id. Default is "div.article_info h3" for naver news title.}

\item{attr}{if you want to get attribution text, please write down here. Defalt is "title".}
}
\value{
Get character press.
}
\description{
Get naver news press name from link.
}
