% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getContentSummary}
\alias{getContentSummary}
\title{Get Content Summary.}
\usage{
getContentSummary(root)
}
\arguments{
\item{root}{is naver news link.}
}
\value{
Get character summary content.
}
\description{
Get naver news summary from link.
}
