% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLike.R
\name{getLike}
\alias{getLike}
\title{Get like Count}
\usage{
getLike(turl = url)
}
\arguments{
\item{turl}{like <https://news.naver.com/main/read.nhn?mode=LSD&mid=shm&sid1=100&oid=056&aid=0010335895>.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Get naver news like Count
}
\examples{
\dontrun{
  getLike("https://n.news.naver.com/mnews/article/214/0001195110?sid=103")
}
}
