% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMainCategory.R
\name{getMainCategory}
\alias{getMainCategory}
\title{Get News Main Categories}
\usage{
getMainCategory()
}
\value{
Get data.frame(chr:cate_name, chr:sid1).
}
\description{
Get naver news main category names and ids recently.
}
