% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaxPageNum.R
\name{getMaxPageNum}
\alias{getMaxPageNum}
\alias{news_max_page_num}
\title{Get Max Page Number}
\usage{
getMaxPageNum(turl, max = 100)

news_max_page_num(turl, max = 100)
}
\arguments{
\item{turl}{is target url include sid1, sid2, date like below.
<http://news.naver.com/main/list.nhn?sid2=265&sid1=100&mid=shm&mode=LS2D&date=20161102>}

\item{max}{is also interval to try max page number is numeric. Default is 100.}
}
\value{
Get numeric
}
\description{
Get Max Page Number
}
\examples{
\dontrun{
  getMaxPageNum("https://news.naver.com/main/list.naver?mode=LS2D&mid=shm&sid1=103&sid2=376")
  }
}
\keyword{internal}
