% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNewsTrand.R
\name{getNewsTrand}
\alias{getNewsTrand}
\title{Get Query news trand by date.}
\usage{
getNewsTrand(query, startDate, endDate, onlyPaper = FALSE)
}
\arguments{
\item{query}{requred.}

\item{startDate}{requred form YYYY-MM-DD.}

\item{endDate}{requred form YYYY-MM-DD.}

\item{onlyPaper}{Default is False means all count of internet news.}
}
\value{
Get data.frame(date, cnt).
}
\description{
Get number of query volume in naver news.
}
