% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPageNum.R
\name{getPageNum}
\alias{getPageNum}
\title{Get urls last page number.}
\usage{
getPageNum(url)
}
\arguments{
\item{target}{urls getting function getCategoryUrl()}
}
\value{
Get page number to end
}
\description{
Main category urls are written manually like blow.
}

