% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSid1.R
\name{getSid1}
\alias{getSid1}
\title{Get Main Category}
\usage{
getSid1()
}
\value{
Get data.frame(chr:cate_name, chr:sid1).
}
\description{
Get naver news main category names and urls recently.
There are 7 categories in never news.
1: Politics, 2: Economics, 3: Social, 4: Living / Culture, 5: World, 6: IT / science, 7: Opinion
}

