% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrlList.R
\name{getUrlList}
\alias{getUrlList}
\alias{news_urls_from_list}
\title{Get Url List By Category}
\usage{
getUrlList(turl, col = c("titles", "links"))

news_urls_from_list(turl, col = c("titles", "links"))
}
\arguments{
\item{turl}{is target url naver news.}

\item{col}{is what you want to get from news. Defualt is all.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Get naver news titles and links from target url.
}
\examples{
 \dontrun{
  getUrlList("https://news.naver.com/main/list.naver?mode=LS2D&mid=shm&sid1=103&sid2=376")
  }
}
\keyword{internal}
