% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrlList.R
\name{getUrlListByQuery}
\alias{getUrlListByQuery}
\title{Get naver news titles and links from target url.}
\usage{
getUrlListByQuery(url = url)
}
\arguments{
\item{url}{is target url naver news.}
}
\value{
Get data.frame(news_title, news_links).
}
\description{
Get naver news titles and links from target url.
}

