% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVideoUrl.R
\name{getVideoUrl}
\alias{getVideoUrl}
\title{Get video clip download url in news}
\usage{
getVideoUrl(turl = url)
}
\arguments{
\item{turl}{like <https://news.naver.com/main/read.nhn?mode=LSD&mid=shm&sid1=100&oid=056&aid=0010335895>.}
}
\value{
Get character url.
}
\description{
Get naver news video url
}
\examples{
 \donttest{
  print(video_url_ex)
  getVideoUrl(video_url_ex)
  }
}
