% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchNews.R
\name{html2text}
\alias{html2text}
\title{Change html formatted text to plain text}
\usage{
html2text(x)
}
\arguments{
\item{x}{character to be converted from html format to plain text}
}
\value{
plain text.
}
\description{
Change html formatted text to plain text
}
