% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N2H4-package.R
\docType{package}
\name{N2H4-package}
\alias{N2H4}
\alias{N2H4-package}
\title{N2H4: Handling methods for naver news text crawling}
\description{
Provides some utils to get korean text sample from news articles in
    naver which is popular news portal servise in korea.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/forkonlp/N2H4}
  \item Report bugs at \url{http://github.com/forkonlp/N2H4/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
