% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{query_list_url_ex}
\alias{query_list_url_ex}
\title{Query Result Page Url Example}
\format{example url character}
\usage{
query_list_url_ex
}
\description{
Query Result Page Url Example
}
\keyword{datasets}
