% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchNews.R
\name{searchNews}
\alias{searchNews}
\title{Search Naver News}
\usage{
searchNews(client_id, client_secret, term, start = 1, display = 10,
  sort = c("sim", "date"))
}
\arguments{
\item{client_id}{character. provided by Naver after registering application here : 'https://developers.naver.com'}

\item{client_secret}{character.Refers to 'client_secret' provided by naver when you register application.}

\item{term}{character. Search keyword}

\item{start}{integer. Default value is 1 and max possible value is 1000. Xth number of article to start extracting from.}

\item{display}{integer, Default value is 10, max possible value is 100. The number of articles to extract.}

\item{sort}{character. sort search result by 'sim':similarity or 'date':date.}
}
\value{
Get data.frame.
}
\description{
Search Naver News using Naver news search API
}
