% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUrl.R
\name{setUrlQuery}
\alias{setUrlQuery}
\title{Get naver news urls from query.}
\usage{
setUrlQuery(query = "", targetDate = today() - 1)
}
\arguments{
\item{query}{is target keyword.}

\item{targetDate}{is one date to get news like "2016-01-01". Default date is yesterday.}
}
\value{
Get data.frame(cate_name, cate_sub, cate_url, last_page_num).
}
\description{
Get naver news urls from query.
}

