% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComment.R
\name{getAllComment}
\alias{getAllComment}
\title{Get All Comment}
\usage{
getAllComment(turl = url, ...)
}
\arguments{
\item{turl}{character. News article on 'Naver' such as <http://news.naver.com/main/read.nhn?mode=LSD&mid=shm&sid1=100&oid=056&aid=0010335895>. News articl url that is not on Naver.com domain will generate an error.}

\item{...}{parameter in getComment function.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Get all comments from the provided news article url on naver
}
\details{
Works just like getComment, but this function executed in a fashion where it finds and extracts all comments from the given url.
}
\examples{
\donttest{
  print(news_url_ex)
  getAllComment(news_url_ex)
  }
}
