% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCategoryUrl.R
\name{getCategoryUrl}
\alias{getCategoryUrl}
\title{Get Category Url}
\usage{
getCategoryUrl(select = c(1, 2, 3, 4, 5, 6))
}
\arguments{
\item{select}{from 1 to 6 numeric values which mean categories.}
}
\value{
Get data.frame(cate_name, cate_sub, cate_url).
}
\description{
Get naver news sub_category names and urls recently.
Main category urls are written manually like blow.
cate_code<-c(100,101,102,103,104,105)
There are 6 categories in never news.
1: Politics, 2: Economics, 3: Social, 4: Living / Culture, 5: World, 6: IT / science
}

