% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComment.R
\name{getComment}
\alias{getComment}
\title{Get Comment}
\usage{
getComment(turl, count = 10, type = c("df", "list"))
}
\arguments{
\item{turl}{like <https://n.news.naver.com/mnews/article/023/0003712918>.}

\item{count}{is a number of comments. Defualt is 10. "all" works to get all comments.}

\item{type}{type return df or list. Defualt is df. df return part of data not all.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Get naver news comments.
if you want to get data only comment, enter command like below.
getComment(url)$result$commentList[[1]]
}
\examples{
\dontrun{
  getComment("https://n.news.naver.com/mnews/article/421/0002484966?sid=100")
}
}
