% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComment.R
\name{getComment}
\alias{getComment}
\title{Get Comment}
\usage{
getComment(turl = url, pageSize = 10, page = 1, sort = c("favorite",
  "reply", "old", "new"))
}
\arguments{
\item{turl}{like 'http://news.naver.com/main/read.nhn?mode=LSD&mid=shm&sid1=100&oid=056&aid=0010335895'.}

\item{pageSize}{is number of comment. defult is 10.}

\item{page}{is defult is 1.}

\item{sort}{you can select favorite, reply, old, new. favorite is defult.}
}
\value{
Get data.frame.
}
\description{
Get naver news comments
if you want to get data only comment, enter commend like below.
getComment(url)$result$commentList[[1]]
}

