% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComment.R
\name{getComment}
\alias{getComment}
\title{Get Comment}
\usage{
getComment(
  turl = url,
  pageSize = 10,
  page = 1,
  sort = c("favorite", "reply", "old", "new", "best"),
  type = c("df", "list")
)
}
\arguments{
\item{turl}{like <https://news.naver.com/main/read.nhn?mode=LSD&mid=shm&sid1=100&oid=056&aid=0010335895>.}

\item{pageSize}{is a number of comments per page. defult is 10. max is 100.}

\item{page}{is defult is 1.}

\item{sort}{you can select favorite, reply, old, new. favorite is defult.}

\item{type}{type return df or list. Defult is df. df return part of data not all.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Get naver news comments
if you want to get data only comment, enter command like below.
getComment(url)$result$commentList[[1]]
}
\examples{
\donttest{
  print(news_url_ex)
  getComment(news_url_ex)
}
}
