% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommentHistory.R
\name{getCommentHistory}
\alias{getCommentHistory}
\title{Get Comment History}
\usage{
getCommentHistory(
  turl,
  commentNo,
  pageSize = 10,
  page = 1,
  type = c("df", "list")
)
}
\arguments{
\item{turl}{character. News article on 'Naver' such as
<https://n.news.naver.com/mnews/article/001/0009205077?sid=102>.
News articl url that is not on Naver.com domain will generate an error.}

\item{commentNo}{Parent Comment No.}

\item{pageSize}{is a number of comments per page. defult is 10. max is 100.}

\item{page}{is defult is 1.}

\item{type}{type return df or list. Defult is df. df return part of data not all.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Get naver news comments on user histories.
}
\examples{
\dontrun{
  getComment("https://n.news.naver.com/mnews/article/421/0002484966?sid=100")
}
}
