% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getContent}
\alias{getContent}
\alias{news_content}
\title{Get Content}
\usage{
getContent(
  turl,
  col = c("url", "original_url", "section", "datetime", "edittime", "press", "reporter",
    "title", "body")
)

news_content(
  turl,
  col = c("url", "original_url", "section", "datetime", "edittime", "press", "reporter",
    "title", "body")
)
}
\arguments{
\item{turl}{is naver news link.}

\item{col}{is what you want to get from news. Defualt is all.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Get naver news content from links.
}
\examples{
\dontrun{
  getContent("https://n.news.naver.com/mnews/article/214/0001195110?sid=103")
  }
}
