% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getContent}
\alias{getContent}
\title{Get Content}
\usage{
getContent(url, col = c("url", "datetime", "press", "title", "body"),
  async = FALSE, ...)
}
\arguments{
\item{url}{is naver news link.}

\item{col}{is what you want to get from news. Defualt is all.}

\item{async}{async crawling if it is TRUE. Defualt is FALSE.}

\item{...}{you can use child function params like title_node_info.}
}
\value{
Get data.frame(url,datetime,press,title,body).
}
\description{
Get naver news content from links.
}
