% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getContent}
\alias{getContent}
\title{Get naver news content from links.}
\usage{
getContent(newslists = AllurlList)
}
\arguments{
\item{newslists}{is target links getting from getAllUrlList function.}
}
\value{
Get data.frame(cate_name, cate_sub, cate_url).
}
\description{
Get naver news content from links.
}

