% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getContent}
\alias{getContent}
\title{Get Content}
\usage{
getContent(
  turl,
  col = c("url", "datetime", "edittime", "press", "title", "body"),
  try_cnt = 3
)
}
\arguments{
\item{turl}{is naver news link.}

\item{col}{is what you want to get from news. Defualt is all.}

\item{try_cnt}{is how many you want to try again if error. Default is 3.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Get naver news content from links.
}
\examples{
\donttest{
  print(news_url_ex)
  getContent(news_url_ex)
  }
}
