% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getContentPress}
\alias{getContentPress}
\title{Get Content Press name.}
\usage{
getContentPress(
  html_obj,
  press_node_info = "div.article_header div a img",
  press_attr = "title"
)
}
\arguments{
\item{html_obj}{"xml_document" "xml_node" using read_html function.}

\item{press_node_info}{Information about node names like tag with class or id. Default is "div.article_info h3" for naver news title.}

\item{press_attr}{if you want to get attribution text, please write down here. Defalt is "title".}
}
\value{
Get character press.
}
\description{
Get naver news press name from link.
}
\examples{
\donttest{
  print(news_url_ex)
  hobj <- xml2::read_html(news_url_ex)
  getContentPress(hobj)
  }
}
