% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContent.R
\name{getContentTitle}
\alias{getContentTitle}
\title{Get Content Title}
\usage{
getContentTitle(
  html_obj,
  title_node_info = "div.article_info h3",
  title_attr = ""
)
}
\arguments{
\item{html_obj}{"xml_document" "xml_node" using read_html function.}

\item{title_node_info}{Information about node names like tag with class or id. Default is "div.article_info h3" for naver news title.}

\item{title_attr}{if you want to get attribution text, please write down here.}
}
\value{
Get character title.
}
\description{
Get naver news Title from link.
}
\examples{
\donttest{
  print(news_url_ex)
  hobj <- rvest::read_html(news_url_ex)
  getContentTitle(hobj)
  }
}
