% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMainCategory.R
\name{getMainCategory}
\alias{getMainCategory}
\title{Get News Main Categories}
\usage{
getMainCategory()
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Get naver news main category names and ids recently.
}
\examples{
\donttest{
  getMainCategory()
  }
}
