% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaxPageNum.R
\name{getMaxPageNum}
\alias{getMaxPageNum}
\title{Get Max Page Number}
\usage{
getMaxPageNum(url = url, max = 100)
}
\arguments{
\item{url}{is target url include sid1, sid2, date like below.
'http://news.naver.com/main/list.nhn?sid2=265&sid1=100&mid=shm&mode=LS2D&date=20161102'}

\item{max}{is also interval to try max page number is numeric. Default is 100.}
}
\value{
Get numeric
}
\description{
Get Max Page Number
}

