% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNewsTrend.R
\name{getNewsTrend}
\alias{getNewsTrend}
\title{Get Query news trend by date.}
\usage{
getNewsTrend(query, startDate, endDate)
}
\arguments{
\item{query}{requred.}

\item{startDate}{requred form YYYY-MM-DD.}

\item{endDate}{requred form YYYY-MM-DD.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Get number of query volume in naver news. Params depend on getQueryUrl function.
}
\examples{
\donttest{
  getNewsTrend("endgame", "2019-03-03", "2019-03-04")
  }
}
