% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQueryUrl.R
\name{getQueryUrl}
\alias{getQueryUrl}
\title{Get Query page url}
\usage{
getQueryUrl(
  query,
  startDate = as.Date(Sys.time()) - 3,
  endDate = as.Date(Sys.time())
)
}
\arguments{
\item{query}{requred.}

\item{startDate}{Dfault is 3 days before today.}

\item{endDate}{Default is today.}
}
\value{
url.
}
\description{
Get naver news query page url without pageNum.
}
\examples{
\donttest{
  getQueryUrl("endgame")
  }
}
