% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSubCategory.R
\name{getSubCategory}
\alias{getSubCategory}
\title{Get News Sub Categories}
\usage{
getSubCategory(sid1 = 100, onlySid2 = TRUE)
}
\arguments{
\item{sid1}{Main category id in naver news url. Only 1 value is passible. Default is 100 means Politics.}

\item{onlySid2}{sid2 is sub category id. some sub categories don't have id. If TRUE, functions return data.frame(chr:sub_cate_naem, char:sid2). Defaults is TRUE.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Get naver news sub category names and urls recently.
}
\examples{
\donttest{
  getSubCategory(100)
  getSubCategory(100, FALSE)
  }
}
