% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrlListByCategory.R
\name{getUrlListByCategory}
\alias{getUrlListByCategory}
\title{Get Url List By Category}
\usage{
getUrlListByCategory(turl = url)
}
\arguments{
\item{url}{is target url naver news.}
}
\value{
Get data.frame(news_title, news_links).
}
\description{
Get naver news titles and links from target url.
}

