% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrlListByQuery.R
\name{getUrlListByQuery}
\alias{getUrlListByQuery}
\title{Get Url List By Query}
\usage{
getUrlListByQuery(turl = url)
}
\arguments{
\item{turl}{is target url naver news.}
}
\value{
Get data.frame(news_title, news_links).
}
\description{
Get naver news(only not other sites links) titles and links from target url.
}

