% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUrlListByQuery.R
\name{getUrlListByQuery}
\alias{getUrlListByQuery}
\title{Get Url List By Query}
\usage{
getUrlListByQuery(turl = url)
}
\arguments{
\item{turl}{is target url naver news.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Get naver news(only not other sites links) titles and links from target url.
}
\examples{
 \donttest{
  print(query_list_url_ex)
  getUrlListByQuery(query_list_url_ex)
  }
}
