% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUrl.R
\name{setUrlByCategory}
\alias{setUrlByCategory}
\title{Get naver news urls from category.}
\usage{
setUrlByCategory(select = select, targetDate = today() - 1)
}
\arguments{
\item{select}{is target categories. Default is all}

\item{targetDate}{is one date to get news like "2016-01-01". Default date is yesterday.}
}
\value{
Get data.frame(cate_name, cate_sub, cate_url, last_page_num).
}
\description{
There are 6 categories in never news.
1: Politics, 2: Economics, 3: Social, 4: Living / Culture, 5: World, 6: IT / science
}

