% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUrls.R
\name{setUrls}
\alias{setUrls}
\title{Set url for crawling}
\usage{
setUrls(
  sid1_vec,
  sid2_vec,
  strDate,
  endDate,
  page_vec = NA,
  return_type = c("list", "df")
)
}
\arguments{
\item{sid1_vec}{is news code in naver news url}

\item{sid2_vec}{is news code in naver news url.}

\item{strDate}{target date of start.}

\item{endDate}{target date of end.}

\item{page_vec}{pageNum default is NA.}

\item{return_type}{list or data.frame. default is list.}
}
\value{
Get data.frame(sid1,sid2,date,pageNum,pageUrl) or list(sid1,sid2,date,pageNum,pageUrl)
}
\description{
Set naver news links with sid, date, etc.
sid1, sid2, page can use vectors.
sid1, sid2, start Date, end Date is requred.
}
\examples{
 \donttest{
  setUrls(105, 227, "20180101", "20180102")
  }
}
